Imported Jammo Character and Behaviour Package
Also made prefab for guard with needed inspector stuff. Also made a behavior graph for the guard that is untouched. There is an issue with the material I think where it is not working properly (guard is just pink on my end). I am unsure the solution and will find a fix later maybe.